Working branch

+++ XYZCompany/Services/EmployeeService.cs
@@ -62,6 +62,26 @@ namespace XYZCompany.Services
 
         public async Task<EmployeeResponse> Create(EmployeeRequest request)
         {
+            if (string.IsNullOrEmpty(request.FirstName))
+            {
+                throw new Exception("FirstName cannot be empty.");
+            }
+
+            if (string.IsNullOrEmpty(request.LastName))
+            {
+                throw new Exception("LastName cannot be empty.");
+            }
+
+            if (request.BirthDate == default)
+            {
+                throw new Exception("BirthDate cannot be empty.");
+            }
+
+            if (request.TitleId == default)
+            {
+                throw new Exception("TitleId cannot be empty.");
+            }
+
             var employee = new Employee
             {
                 FirstName = request.FirstName,
@@ -81,6 +101,27 @@ namespace XYZCompany.Services
         {
             var employee = await _employeeRepository.Get(id);
 
+            if(string.IsNullOrEmpty(request.FirstName))
+            {
+                throw new Exception("FirstName cannot be empty.");
+            }
+
+            if (string.IsNullOrEmpty(request.LastName))
+            {
+                throw new Exception("LastName cannot be empty.");
+            }
+
+            if(request.BirthDate == default)
+            {
+                throw new Exception("BirthDate cannot be empty.");
+            }
+
+            if (request.TitleId == default)
+            {
+                throw new Exception("TitleId cannot be empty.");
+            }
+
+
             employee.FirstName = request.FirstName;
             employee.LastName = request.LastName;
             employee.BirthDate = request.BirthDate;
@@ -95,6 +136,13 @@ namespace XYZCompany.Services
 
         public async Task Delete(Guid id)
         {
+            var employeeToDelete = await _employeeRepository.Get(id);
+
+            if(employeeToDelete == null) 
+            {
+                throw new Exception($"Record with Id {id} not found.");
+            }
+
             await _employeeRepository.Delete(id);
         }
     }
